services/auth-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0
pydantic[email]>=2.4.0
